refactor: Comment 수정 요청 픽스처 분리
